+++ languagetool-standalone/src/test/java/org/languagetool/language/LanguageIdentifierTest.java
@@ -92,6 +92,12 @@ public class LanguageIdentifierTest {
             "marka ay dhacdo dhibaato la xiriirta dulimaad.");
   }
 
+  @Test
+  public void testIgnoreSignature() {
+    langAssert("de", "Das ist ein deutscher Text\n-- \nBut this is an English text in the signature, and it's much longer than the original text.");
+    langAssert("en", "This is an English text.\n-- \nDas ist ein deutscher Text in der Signatur, der länger ist als der Haupttext.");
+  }
+  
   private void langAssert(String expectedLangCode, String text) {
     langAssert(expectedLangCode, text, identifier);
   }