+++ languagetool-standalone/CHANGES.md
@@ -19,6 +19,8 @@
 #### General
   * Prepared support for AIX. See https://github.com/MartinKallinger/hunspell-aix
     for the required libraries
+  * Email signatures are now ignored for language detection as long as they are
+    separated from the main text with `\n-- \n` 
 
 #### HTTP API / LT server
   * The server can now accept JSON as the `data` parameter that describes